[Android] Fix for tab switcher mode in stack layout
Stack layout in not taken into consideration in Chromium anymore and
unintendedly was broken by the following change:

https://chromium.googlesource.com/chromium/src/+/dc3038ca3e29694ab91dfc1832ff7b337489fcfa

Fix tab switcher button not working in a11y mode

This patch is a temporary fix while grid tab switcher is in the process
of being launched. Previously, accessibility mode was one of the cases
where the grid was not created. This poses a problem since the a11y
mode can change without the activity restarting. This patch removes the
a11y check for creating the grid and adds a special case to the
overview showing logic in ChromeTabbedActivity to handle this.

Bug: 1200727